some changes made to responsive design

+++ frontend/src/components/WeatherForecast.jsx
@@ -24,7 +24,7 @@ export default function WeatherForecast({coords,mode}){
                 <div className="grid grid-cols-5 mt-10 max-md:grid-cols-1">
                     {forecastData.length!==0 && forecastData.map((item,index)=>(
                         <div className="bg-stone-200 bg-opacity-45 mx-auto w-[80%] rounded-lg flex flex-col justify-center items-center max-md:mt-4" key={index}>
-                            <p>{Intl.DateTimeFormat('en-US',{weekday:'long'}).format(new Date(item[0]))}</p>
+                            <p>{Intl.DateTimeFormat('en-US',{weekday:'short'}).format(new Date(item[0]))}</p>
                             <img src={`https://openweathermap.org/img/wn/${item[4]}@2x.png`} alt={item[3]} />
                             <p>{item[3]}</p>
                             <p><span>{item[1]}&deg;</span> / <span>{item[2]}&deg;</span></p>
